// src/app/modules/academics/student-information-system/model/student-request-view.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AttachmentViewModel } from './attachment-view-model';


/**
 * StudentRequestView
 */
export interface StudentRequestView { 
    /**
     * Gets or Sets StudentId
     */
    studentId?: string;
    /**
     * Gets or Sets RequestNote
     */
    requestNote?: string;
    /**
     * Gets or Sets RequestTypeId
     */
    requestTypeId?: number;
    /**
     * Gets or sets Attachment
     */
    attachment?: AttachmentViewModel;
}