// src/app/modules/academics/student-information-system/model/student-action-view.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * StudentAction
 */
export interface StudentActionView { 
    /**
     * Gets or set
     */
    details?: string;
    /**
     * Gets or set
     */
    studentId?: string;
    /**
     * Gets or set
     */
    studentActionTypeId?: number;
}