// src/app/modules/academics/student-information-system/model/siblingStudentView.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * SiblingStudentView
 */
export interface SiblingStudentView { 
    /**
     * Gets or sets
     */
    studentId1?: string;
    /**
     * Gets or sets
     */
    studentId2?: string;
    /**
     * Gets or sets
     */
    studentId3?: string;
    /**
     * Gets or sets
     */
    studentId4?: string;
    /**
     * Gets or sets
     */
    studentId5?: string;
}